requirements.txt
absl-py==2.1.0
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
astunparse==1.6.3
audioread==3.0.1
backoff==2.2.1
bcrypt==4.2.0
billiard==4.2.1
blinker==1.8.2
build==1.2.2.post1
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.15
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
decorator==5.1.1
Deprecated==1.2.14
durationpy==0.9
fastapi==0.115.3
filelock==3.16.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fsspec==2024.10.0
gast==0.6.0
google-auth==2.35.0
google-pasta==0.2.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.6.0
kombu==5.4.2
kubernetes==31.0.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.8
networkx==3.4.2
nltk==3.9.1
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
onnxruntime==1.19.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
opt_einsum==3.4.0
optree==0.13.0
orjson==3.10.10
overrides==7.7.0
packaging==24.1
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
posthog==3.7.0
prompt_toolkit==3.0.48
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.3
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
SpeechRecognition==3.11.0
SQLAlchemy==2.0.36
starlette==0.41.0
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
tokenizers==0.20.1
torch==2.5.0
tqdm==4.66.5
transformers==4.46.0
triton==3.1.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.6
wrapt==1.16.0
zipp==3.20.2
